Override NIX_PATH when running tinc
Internally, the `tinc` command calls:

  nix-shell -p 'haskellPackages.ghcWithPackages (p: [ p.cabal-install ])'
            --pure --run 'cabal ....'

Using the haskellPackages global in this way completely bypasses whichever
haskell package set was given to tincify. To remedy this, we need to serialise
the given package set into a file, and read it back in tinc's nix-shell call.

Since we know that only cabal-install is needed, we build this derivation
up-front. We then construct a directory containing a .nix file which only
provides a haskellPackages.ghcWithPackages attribute, which is hard-coded to
return our pre-built derivation. We override NIX_PATH to use this directory as
<nixpkgs>, so that nix-shell will get the right environment.