Added shouldFail
Useful for tests, we can say 'shouldFail myDrv', and get back a derivation which
succeeds (outputting a message to "$out") when myDrv's 'builder' fails, and
fails when myDrv's 'builder' succeeds.

This can be especially useful if we're implementing some workaround. For
example, if 'somePkg' doesn't work with the latest GHC, we might override it to
use a previous GHC. In such cases, we might also add 'shouldFail originalPkg' as
a dependency, which will trigger a build failure when our workaround is no
longer needed.