Add more helpers:
`forceBuilds [ a b c ]` will build (at eval time) the derivations `a`, `b` and
`c`, then return `true`; if a derivation fails to build, evaluation will abort
as normal. This is useful for assertions, although keep in mind that it's
preferable to do things at build time rather than eval time where possible,
since expressions are evaluated more than they're built (since every build
requires an evaluation, but not all evaluations are for building).

`withDeps [ a b c ] d` will add the derivations `a`, `b` and `c` to the
dependencies of `d`. The result will behave the same as `d`, except that when we
build it the `a`, `b` and `c` derivations will be built first (just like any
other dependencies); if they build successfully then the result will be
equivalent to `d` (i.e. it will build successfully iff `d` does, and provides
the same outputs); if any of them fails, the build will be aborted.

`withDeps` is useful for adding extra tests to an existing derivation.