Fix quoting of variables in wrap
makeWrapper makes a half-arsed attempt at quoting inputs; this fails to properly
escape some inputs (e.g. anything containing $), but it also prevents us from
properly escaping the inputs we pass in (we end up with the quotation marks and
backslashes of our escaping appearing verbatim in the string!).

This fixes the issue, using a combination of double-escaping and working around
the existing, broken escaping.

We include a bunch of tests to ensure the results work as expected.

Note that we currently only do this for setting variables; we don't do it for
things like paths or filenames.

There are upstream changes in nixpkgs which may break this implementation after
an upgrade, but that's OK since the tests will tell us if/when that causes any
problems.