Don't force wrap dependencies at eval time
Previously, wrap was splicing the given paths and variables into the build
command, which required building any given derivations at eval time in order to
construct the buildCommand string. This is less than ideal!

We now add a layer of indirection: paths and variables are added to the
environment of buildCommand, rather than spliced in; these values are added
"as-is", so that derivations are left untouched.

This requires us to perform all of our manipulation (escaping, etc.) in Bash, at
build time, rather than in Nix like we were doing before. This is a little ugly,
and requires a few levels of indirection, but the tests are all
passing (including new ones which verify that the paths and vars aren't forced,
by using some broken derivations which would cause Nix to bail if they ever did
get forced)