Make stable.nix less "dynamic"
We're having a problem with circular imports, which may be caused by
stable. Even if it's not, this new definition seems clearer. Basically, we were
calculating a bunch of attrsets and folding them together. Now we're explicitly
writing down the attribute names (well, we also transform some with mapAttrs',
but meh).

Since config.nix merges together the results from these top-level files, they're
going to be forced every time we evaluate anything (at least, the
names/structure are), so it's probably a bad idea to determine that structure
"dynamically", where there's a chance we'll get stuck in a loop.