Add hackagePackageNames
We look through the our stable Hackage index and return all of the package names
that we find.

This is useful for telling whether a package is a special snowflake that won't
be found by Cabal's dependency resolver.

We include 2 versions: hackagePackageNamesDrv is a derivation that does the
extraction; hackagePackageNames imports the result of this.

We will probably want the latter most of the time, but it requires building the
hackage index which takes a long time; if we're only using these values inside a
build script then it might make sense to use the latter, as it won't slow down
evaluation.