Remove layer of indirection in tip-lib
We were patching the source, calling 'nix-store --add', writing the resulting
path to our output, importing our output and using that (string) path as the
'src' of the subsequent derivation. This is too much for Nix's garbage
collector, which can't see that the resulting package depends on the thing we
added using 'nix-store --add'.

Since this is unnecessary anyway (oh the naivety of youth!), I've dropped the
'nix-store --add' in favour of just copying to '$out', and rather than importing
the result to get a string, we just depend on the patching derivation directly.

This should prevent problems with paths getting garbage collected, and fix
mysterious "/nix/store/....-tip not found" errors.