Force all overrides to use stable nixpkgs version instead of <nixpkgs>
Our custom packages and overrides are prone to breaking, which is annoying since
Nix is meant to aid isolation and reproducibility.

One of the main reasons is that we're extending the <nixpkgs> set, which may
vary across systems. To remove this ambiguity, we now use a fixed, known nixpkgs
set (currently nixpkgs1603) as the basis for all of our packages/overrides.

We actually provide packages/overrides for various nixpkgs sets (16.03, 16.09,
etc.) including the system's <nixpkgs> (which we call "unstable"). These are
all added to an attribute called 'customised'.

For each custom thing 'foo' there will be a 'customised.nixpkgs1603.foo', a
'customised.nixpkgs1609.foo', etc. as well as a
'customised.stable.foo' ('stable' is currently a synonym for 'nixpkgs1603') and
'customised.unstable.foo' (which is based on <nixpkgs>).

The top-level 'foo' will be taken from 'customised.stable', rather than
'customised.unstable' (which would be equivalent to the old, unreliable
behaviour).