Use latestGit rather than fetchgit
On further thought, we shouldn't be replacing fetchgit, since that will affect
derivations inside nixpkgs which provide hashes for their expected version of
fetchgit.

Instead, we should give the constant version of fetchgit a new name, and use
that for all of our own things. Since latestGit already does this, with the
(questionable) bonus of picking stable/unstable for us, we use this instead.