Use strings containing paths to scripts, rather than derivations
Some systemd attributes apparently need to be strings containing the commands to
execute, so we're not allowed to put a script derivation there. This can be
solved by using 'toString' to get the path of the resulting script, since that
is a valid command which will run the script as we want.