Modify release.nix to be more lazy
We were checking to see if each custom entry in nixpkgs was a derivation or not,
and discarding those which aren't (since Hydra can only build derivations, or
sets of derivations).

The problem is, some of these entries are generated in quite elaborate ways,
e.g. fetching some external repo, running some builders, importing the results,
etc. so this could take a while just to evaluate.

Since they're our overrides, we should know which ones are derivations and which
aren't, so we might as well just write them down.