Remove AWS/S3 from Hydra
aws-sdk-cpp doesn't build on 32bit machines, since it needs more than 4GB for
GCC, which isn't addressable with 32bits.

We patch the source to remove all references to AWS, which is just the S3
functionality which we don't use anyway.

We also need to do other fixups, like removing references to the build directory
from the final wrapper scripts.