Make haskellRelease more robust
In particular, take into account the fact that some arguments must come from the
nixpkgs set, but we can only tell this by attempting to call the package
function and seeing what it does with the arguments.

We also force an old version of zlib to be used, since newer ones break the zlib
Haskell package's test suite.