Fix importing with non-default defaultVersion arguments
We should be able to import ./. { defaultVersion = ...; } and have a nicely
overridden nixpkgs set for 16.03, 16.09, 17.03, 17.09 or 18.03. Yet there was a
problem with importing 16.* on a nixos 18.03 system. I'm not sure what the cause
is, but it's probably something to do with the "overlays" mechanism.

In any case, we now *always* import the 'stable' version, use that to get a
"repo" for the given version (e.g. if defaultVersion == nixpkgs1609 then we get
repo1609). We import *that* with our overrides and the given defaultVersion.

This at least seems to avoid the set/function error we were getting before.