Use 'nixpkgsVersion' rather than 'version' to avoid breaking fontconfig
Some nixpkgs packages are defined like:

    { fetchurl, foo, bar, sha256 ? "abc", stdenv, version ? "1.2.3" }:
    stdenv.mkDerivation {
      inherit version;
      name = "baz-${version}";
      src  = fetchurl {
        inherit sha256;
        url = "http://example.org/baz-${version}.tar.gz";
      };
    }

This is the case for 'fontconfig-penultimate', for example. Since we're adding a
'version' attribute to nixpkgs, it was getting propagated to such functions and
causing them to use non-existent package versions like 'nixpkgs1803'.

This commit changes our attribute name to 'nixpkgsVersion' which should
hopefully not conflict with anything.