Perform keyhole surgery on asv-nix to ensure Nix dependency gets overridden
asv-nix depends on a test which checks that we can benchmark an example project.
Since asv-nix uses Nix, that test's builder relies on 'withNix' to augment the
environment with appropriate variables, etc. for running Nix commands. Since we
want asv-nix to stand alone, not as part of nix-config, it doesn't rely on
withNix being available in the packageOverrides. Instead, it fetches a
nix-config git commit, imports it and takes withNix from there. This has the
effect of pinning that Nix package, or at least avoiding any overrides we might
have (specifically, our disabling of libuv's flaky test suite).

I think it's fine to leave asv-nix as it is for the time being: if a standalone
use of it runs into problems then we can cross that bridge when we come to it;
in particular, we might have the callers of 'withNix' pass in dependencies like
'nix', either directly as a function call or indirectly by e.g. augmenting their
own dependencies.