nix-daemon-tunnel improvements
Rather than sleeping for 1 second and assuming the socket got created, we now
loop 60 times, sleeping 1 second per iteration. We break out of the loop early
if the socket file is detected.

We also add systemd dependencies on sshd.service and nix-daemon.service, since
we need both of those running to work.

Also fixed the cleanup to not complain about trying to delete a missing file.