Bump stableVersion to 1909
Time to bite the bullet and upgrade. Looks like we can throw away a bunch of
workarounds (e.g. for broken Qt) which is nice. Unfortunately it looks like
haskellPackages is broken (hspec-core tests fail, and that's a dependency of
pretty much everything). I've switched Haskell executables to using haskell-nix
instead, in anticipation.