Disable kernel modules with failing assertions
Some kernel modules are asserting that the version is above/below a certain
number, which is fair enough. Unfortunately, some changes in NixOS/pkgs 19.09
has caused all of these modules to start being evaluated, which is making our
evaluation fail.

I tried disabling them, but the only option which seems to work is patching the
Nixpkgs source to make them 'null'.