Rearrange which metapackage things appear in
I want to share parts of this config with a macOS machine, so we can't just
assume that everything in metaPackages will be installed (via 'allPkgs').

In particular, there are some things which just aren't appropriate on macOS,
like Linux-specific tools. We should also be more strict about what the purpose
of each meta-package is: e.g. should we put get_iplayer in 'netCli' since it's a
downloading tool, or 'mediaCli' since it's for getting videos? In this case I
think it's better to focus on the intended usage (e.g. watching videos), in
which case netCli should be limited to things like packet sniffers, download
managers, etc. since those are intended for playing with networks.

Importantly, this lets us separate 'work' from 'pleasure' a bit more, since a
work machine should have netCli but not necessarily mediaCli.