overlays: move logic from override.nix to overlays.nix
This consolidates the logic, so we can reference any of the overrides/ files via
attrs of the set returned by overlays.nix. I've kept the override.nix file for
backwards compatibility, but it shouldn't really be used now: it just merges all
of the attributes returned by overlays.nix into one big overlay.