warbo.nix: add warbo.is-nixos option (NixOS module turns it on)
This lets our Home Manager module tell whether it's being used standalone (like
on manjaro-arm) or as part of a NixOS configuration (like on nixos-amd64). This
works by defaulting to 'false', and the NixOS module sets it to 'true'.

This is useful if there are different ways to achieve something. In this case,
we'd prefer enabling direnv via NixOS, since it has an extra loadInNixShell
option. Yet we want Home Manager to enable it if we're not on NixOS.