warbo.nix: splice bashrc instead of sourcing
Splicing ensures that its '<is non-interactive> && return' check returns from
the whole .bashrc file, and since it's an '&&' the return code is success.

Sourcing was causing the return to only jump back into the rest of
the (HM-generated) .bashrc. That has its own check '<is interactive> || return',
which returns from the whole file; but since it's an '||' the return code is
failure. That was causing any script that (a) uses 'set -e' and (b) sourced our
.bashrc (to get useful env vars) was aborting.