talecast: fix for finished downloads
TaleCast itself uses 'fs::rename' to move finished downloads from the
'partial_path' to the 'download_path'. However, that only works if they're both
on the same filesystem!

We want partial downloads to use our SSD at /mnt/internal, but finished
downloads to go on our shared drive pool at /mnt/shared. To do that, we'll need
to use a separate service that rsyncs them across, path-triggered by the
download_path. In fact, there's a complication since download_path may contain
placeholders like '{podname}' for putting the name of the podcast in the path.
To avoid that, we specify the download_path as an extension of a path
literal (with no placeholders) we call fetched_path; then we can use the latter
in our rsyncing service. We leave this configurable, but disable the rsyncing
service if download_path is altered from that known-good default.