talecast: avoid race condition in talecast services
The talecast service creates empty directories in the fetched_path straight away
but that triggers the talecast-move service. If talecast-move sees that there
are no files to move and exits, it will restart immediately since the
directories are still there; systemd will eventually disable it, after too many
restarts. If it instead deletes directories (once everything's been moved out of
them, if they were non-empty) then that prevents the restarting, but it also
causes talecast to fail when it tries to move completed downloads into those
directories.

As a compromise, we now do the following in the talecast service:

 - After the talecast command has finished, we touch a "doneFile".
 - Before the talecast command starts, it deletes the "doneFile" (if it exists)

This way, the "doneFile" does not exist while the talecast service is fetching
podcasts. The talecast-move service is changed as follows:

 - If there is no "doneFile", keep looping through the moving logic. This way,
   we will still spot finished downloads promptly, but systemd won't disable the
   service.
 - If the "doneFile" exists, that means the podcasts aren't being fetched (at
   least, by the talecast service), so the talecast-move service can rmdir each
   subdirectory once all of its contents (if any) have been moved out.

The resulting behaviour is that the talecast service will start (triggered by a
timer), and that will delete the "doneFile", then create the subdirectories in
the download_path; that will start the talecast-move service, which will scan
through those (empty) subdirectories and see there's nothing to move, sleep for
a few seconds then try again, over and over. At the same time, talecast will be
moving completed downloads into those subdirectories, which will be spotted the
next time talecast-move loops around.

Once all of the downloads have finished, the "doneFile" is touched. From that
point on, the talecast-move service will rmdir each directory once it's finished
moving its contents (if any) and its loop will stop. If any subdirectories
remain in the download_path (i.e. the final loop iteration had already passed
them by the time "doneFile" appeared), then talecast-move will be immediately
triggered again, and run one iteration of its loop to process and delete those
subdirectories. The talecast-move service will then stop, since the
download_path is now empty.