nix-backport: take from upstream Nixpkgs, and provide NixOS module
Firstly, upstream Nixpkgs now has Nix 2.28, so use that definition; rather than
trying to fight the flake-based definitions in the upstream Nix repo.

Secondly, importing a separate Nixpkgs can cause problems with cross-compilation
if we don't pass along the localSystem and crossSystem. The easiest way to
access those is inside a NixOS module, so provide one of those. That can also
take care of setting up nix-backport as the value of nix.package.

This now works on RiscV.